リモートリポジトリの変更

+++ src/app/Http/Controllers/LikeController.php
@@ -10,18 +10,22 @@ class LikeController extends Controller
 {
     public function like(Request $request, $item_id)
     {
-        Like::create([
-            'user_id' => $request->user_id,
-            'item_id' => $item_id,
-        ]);
+        if (!empty($request->user_id)) {
+            Like::create([
+                'user_id' => $request->user_id,
+                'item_id' => $item_id,
+            ]);
+        }
         $like_number = Like::where('item_id', $item_id)->count();
         return response()->json([$like_number]);
     }
 
     public function unlike(Request $request, $item_id)
     {
-        $like = Like::where('user_id', $request->user_id)->where('item_id', $item_id)->first();
-        $like->delete();
+        if (!empty($request->user_id)) {
+            $like = Like::where('user_id', $request->user_id)->where('item_id', $item_id)->first();
+            $like->delete();
+        }
         $like_number = Like::where('item_id', $item_id)->count();
         return response()->json([$like_number]);
     }
